+++ CHANGELOG.rst
@@ -4,6 +4,15 @@ CHANGELOG
 6.0.0a13 (unreleased)
 ---------------------
 
+- Breaking API change: Search GET 
+  Search get responds a json with items and items_total like plone rest api
+  [bloodbare]
+
+- Breaking Internal API change: Search
+  Catalog utility search is the public search operation that is parsed and query
+  the internal implementation
+  [bloodbare]
+
 - Fixing WS bugs and redis unsubscription
   [bloodbare]